organized file structure, deleted extraneous imports
