+++ client/src/components/Map/Marker/Marker.js
@@ -1,6 +1,6 @@
 import React from "react";
-import Icons from "../Icon/Icon";
-import MiniPost from './MiniPost'
+import Icons from "../../Icon/Icon";
+import MiniPost from '../MiniPost/MiniPost'
 import styles from "./Marker.module.css";
 
 const defaultIcon = <Icons category={"mapMarker"} />;
