+++ client/src/components/Map/MiniPost/MiniPost.js
