+++ client/src/components/Map/MiniPost/MiniPost.module.css
